pyproject.toml
[project]
name = "voxbot"
version = "0.2.0"
description = "TTS Discord bot using hikari + crescent"
readme = "README.md"
authors = [
    { name = "boonhapus", email = "me@boonhap.us" }
]
requires-python = ">=3.14"
dependencies = [
    "attrs>=26.1.0",
    "cyclopts>=4.10.1",
    "hikari>=2.0.0",
    "hikari-crescent>=5.0.0",
    "hikari-wave>=0.2.0",
    "mistralai>=2.1.3",
    "python-dotenv>=1.2.2",
    "structlog>=25.5.0",
]

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"

[project.scripts]
voxbot = "voxbot.__main__:cli"

[dependency-groups]
dev = [
    "ruff>=0.11.0",
    "ty>=0.0.26",
]